requirements.txt
django<1.8
django-modeltranslation==0.8